vim-patch:8.1.1736: viminfo support is spread out
Problem:    Viminfo support is spread out.
Solution:   Move more viminfo code to viminfo.c. (Yegappan Lakshmanan,
            closes vim/vim#4717)  Reorder code to make most functions static.

https://github.com/vim/vim/commit/c3328169d5566b97a6a6921067017e4369dd7cd6

735aa4c4c89943b26f1d6ba0d3e076002490c09d was the partial port for
the typedefs.
This patch completes the viminfo->shada port.

- get_shada_parameter()
- find_shada_parameter()

Other patches below are N/A.

vim-patch:8.1.1728: wrong place for command line history viminfo support

Problem:    Wrong place for command line history viminfo support.
Solution:   Move it to viminfo.c.

https://github.com/vim/vim/commit/5f32ece459d1f310b1b48b72e07dcd77d3261a76

vim-patch:8.1.1730: wrong place for mark viminfo support

Problem:    Wrong place for mark viminfo support.
Solution:   Move it to viminfo.c. (Yegappan Lakshmanan, closes vim/vim#4716)

https://github.com/vim/vim/commit/1e78e69680a5f52970d9b1ef60710e556b09b8c2
